Web Animations API: Start implementation of Element.animate().
This is the first in a series of patches to implement the Web Animations API (see http://dev.w3.org/fxtf/web-animations/).

Please note that this work is being completed behind a runtime flag.

While the Web Animations model is now used to support CSS Animations and Transitions, there are some important differences between that use and the API:
* CSS Animations and Transitions operate on computed style, and are inherently resolution-dependent mechanisms. Web Animations is style resolution independent.
* CSS Animations and Transitions accept the use of -webkit prefixed properties. Following the principle of deprecating prefixing for new APIs, the Web Animations JS API does not. Note that this *may* require us to gate release of this feature on unprefixing of -webkit-transform.

As an initial patch, this code takes some shortcuts that will eventually be replaced. Most notably, we temporarily make use of StyleResolver for the purpose of parsing CSS Strings into usable values. This will be switched out in the future in favor of an approach that doesn't resolve styles at the same time.

BUG=327564

Review URL: https://codereview.chromium.org/96283002

git-svn-id: svn://svn.chromium.org/blink/trunk@163761 bbb929c8-8fbe-4397-9dbb-9b2b20218538