Fix rendering of select lists on Android.
We have multiple definitions of the WebThemeEngine class in the Blink public/ api. In particular, the android/ version is a clone of the default/ version. Unfortunately, the Android implementation of RenderTheme (RenderThemeChromiumAndroid) *subclasses* RenderThemeChromiumDefault, meaning that it's possible to use the default/ header in some places and the android/ header in others, so we end up using the modified data structure on one side of the call, and the unmodified data structure on the other. 

When I implemented support for the "mock" theme under Aura (for Windows and Linux), I changed one of them (the one in public/platform/default) but not the other. Unhappiness followed as we essentially stomped on memory.

R=darin@chromium.org
BUG=320898

Review URL: https://codereview.chromium.org/111243002

git-svn-id: svn://svn.chromium.org/blink/trunk@163826 bbb929c8-8fbe-4397-9dbb-9b2b20218538