ASSERTION FAILED: !object || object->isTable()
The issue is caused by call sites of
RenderBlock::addChildIgnoringAnonymousColumnBlocks
that basically bypass the RenderTable::addChild logic causing
all sorts of mayhem. The fix I decided on is to fix the previous
RenderBlock function as it's the safest approach to the current
madness.

This change adds an extra ASSERT into RenderObjectChildList to
catch further problem earlier.

BUG=309421

Review URL: https://codereview.chromium.org/111873005

git-svn-id: svn://svn.chromium.org/blink/trunk@164125 bbb929c8-8fbe-4397-9dbb-9b2b20218538