Remove usage of Length(Undefined) as a placeholder in GridLength
This patch simplifies some of the logic in creating and applying
GridLength and GridTrackSize values to RenderStyles by using the
constructors for initialisation instead of deferring to setters.

Neither of createGridTrackSize nor createGridTrackBreadth should
be able to fail, so these are changed to return objects directly
instead of taking an out reference (note that in general failing
in applying properties is incorrect and we should catch these in
the parser).

I've also removed copying the quirk value across to the Lengths,
this doesn't apply to Grid (maybe we can remove the Length quirk
in a later patch, it only appears to be used for margins).

BUG=329365

Review URL: https://codereview.chromium.org/117793003

git-svn-id: svn://svn.chromium.org/blink/trunk@164154 bbb929c8-8fbe-4397-9dbb-9b2b20218538