Speed up convolutions on Windows.
The DirectConvolver represents a significant fraction of the cost of doing convolution because Windows is compiled using the default /fp:precise using the x87 FPU.  This causes the result of each FP operation to be stored to memory and reloaded to preserve the IEEE-specified rounding for floats.

For the convolver, this is not needed, so /fp:fast is used for the DirectConvolver and other support routines that show up as hotspots when running the test case in crbug.com/316740.

For the record, the unpatched Chrome running on a Z600 and Win7 takes 3259 ms on the JS benchmark test in the bug report.  The profile results are

4.570 DirectConvolver::process
2.453 rdft_calc_c
1.352 zvmul                     
1.231 ff_fft_calc_avx           
1.136 vsmul                     
1.133 ff_fft_permute_sse        
0.897 vadd                      
0.768 doFFT                     
0.591 getUpToDateComplex
0.455 OscillatorNode::process        

The routines rdft_calc_c and ff_fft_* are in ffmpeg.

After the patch, the test takes 2460 ms with the following profile results:

2.574 rdft_calc_c               
1.337 DirectConvolver::process  
1.307 ff_fft_calc_avx           
1.202 ff_fft_permute_sse        
1.191 zvmul                     
1.152 vsmul                     
0.895 vadd                      
0.491 OscillatorNode::process   
0.408 doFFT                     
0.230 getUpToDateComplex        

Nice gains on DirectConvolver, getUpToDateComplex, and doFFT.

BUG=316740

Review URL: https://codereview.chromium.org/116823002

git-svn-id: svn://svn.chromium.org/blink/trunk@164242 bbb929c8-8fbe-4397-9dbb-9b2b20218538