Tiny clean-up in applying 'color' in the StyleBuilder
This patch clarifies slightly what goes on when applying the 'color'
property in the StyleBuilder. Note that since the 'color' property
has a default of RenderStyle::initialColor (i.e. black), it will
never get set as 'invalid' on the RenderStyle, so the previous
default_getter argument is redundant.

BUG=331126

Review URL: https://codereview.chromium.org/122603002

git-svn-id: svn://svn.chromium.org/blink/trunk@164389 bbb929c8-8fbe-4397-9dbb-9b2b20218538