Fix WebMediaPlayerClientImpl so it only holds its reference to the MediaSource while opening.
The WebMediaPlayerClientImpl was holding onto its MediaSource reference
much longer than it needed to. Under certain circumstances it can hold
the last reference to the MediaSource, which can lead to crashes when
the MediaSource is destroyed as part of WebMediaPlayerClientImpl
destruction during document teardown.

This change fixes the crashes by only allowing WebMediaPlayerClientImpl
to hold the reference for the mimimal time it needs to. It only needs
to hold the reference while opening the MediaSource. Once the
MediaSource is open, the reference is cleared so it can't cause problems
later.


BUG=330026
TEST=LayoutTests/http/tests/media/media-source/mediasource-gc-after-decode-error-crash.html

Review URL: https://codereview.chromium.org/107933008

git-svn-id: svn://svn.chromium.org/blink/trunk@164455 bbb929c8-8fbe-4397-9dbb-9b2b20218538