Remove unnecessary HTMLMediaSource reference in WebMediaPlayerClientImpl and cleanup load code.
WebMediaPlayerClientImpl was maintaining an unnecessary reference to the
HTMLMediaSource object that the HTMLMediaElement using for the current
load. This change simply removes the reference and moves some load logic
out of WebMediaPlayerClientImpl and into HTMLMediaElement. This brings
is one step closer to removing WebMediaPlayerClientImpl.

Eventhough a reference to the HTMLMediaSource is being removed, this
change should not shorten the object's lifetime because the
HTMLMediaElement and WebMediaPlayerClientImpl should never be holding
the last references to the object.

TESTS= All existing tests still pass.

Review URL: https://codereview.chromium.org/99043006

git-svn-id: svn://svn.chromium.org/blink/trunk@164474 bbb929c8-8fbe-4397-9dbb-9b2b20218538