Remove HTMLTextAreaElement::rendererWillBeDestroyed()
This code was added 6 years ago in https://trac.webkit.org/changeset/13051
and the test that was added with it fast/forms/textarea-setvalue-submit.html
still passes without the code now. It doesn't make any sense that form
controls would need to do this update in the destructor of the renderer,
especially not only <textarea> since none of the other controls do this.

Review URL: https://codereview.chromium.org/123693004

git-svn-id: svn://svn.chromium.org/blink/trunk@164587 bbb929c8-8fbe-4397-9dbb-9b2b20218538