Stopped using GraphicsContext3D for all WebGLRenderingContext GL calls
This CL converts the GraphicsContext3D into a support class for the
WebGraphicsContext3D (at least for this class), which is what almost all calls
are directed through. The remaining calls to GraphicsContext3D are to utility
functions which don't belong on the WebGraphicsContext3D interface.

The conversion done on this class will act as a model for future refactorings,
so if there's anything here that's stylistically or architecturally offensive
it's probably best to take care of it now.

BUG=326382

Review URL: https://codereview.chromium.org/123693002

git-svn-id: svn://svn.chromium.org/blink/trunk@164620 bbb929c8-8fbe-4397-9dbb-9b2b20218538