<body> overflow clip still applies after being removed
The logic after https://codereview.chromium.org/87443002
wouldn't clear the overflow clip bit in every code path.
This would basically leave an overflow clip when the author
removed it. In order to prevent further regression, this change
simplify the nested conditions that made it hard to see when
the overflow clip was actually set or reset.

BUG=326962

Review URL: https://codereview.chromium.org/126063002

git-svn-id: svn://svn.chromium.org/blink/trunk@164675 bbb929c8-8fbe-4397-9dbb-9b2b20218538